Add errors to composition
In order to track errors as we go ahead and traverse the graph of
services we need to transform, we need to track errors in the
composition.

Right before you echo the composition to YAML via stdout, you should
echo to stderr if there are any errors.

For example: if one of your plugins requires a validation step. this
will error out until you fix it (eg: botoenv, placement of a file and
more)